requirements.txt
aiosqlite==0.19.0
anyio==3.6.2
cffi==1.15.1
click==8.1.3
cryptography==40.0.2
databases==0.7.0
dnspython==2.3.0
ecdsa==0.18.0
email-validator==2.0.0.post1
fastapi==0.95.1
fastapy==1.0.3
h11==0.14.0
idna==3.4
passlib==1.7.4
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.7
python-decouple==3.8
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.47
starlette==0.26.1
typing_extensions==4.5.0
uvicorn==0.21.1
